Split imageUtils.js into focused helpers/image/ modules behind a re-export shim
